minor -- removed cast to short from sizeof() operator in this example (this was the convention because it was required for Max externs on 68K processors in the olden-days).
git-svn-id: http://ttblue.googlecode.com/svn/branches/active@606 bc2ad3ca-9a4d-0410-8d14-35da9c29b421

+++ examples/MaxMSP/tt.overdrive~/tt.overdrive~.cpp
@@ -66,7 +66,7 @@ int main(void)
 	TTBlueInit();	
 	common_symbols_init();
 
-	c = class_new("tt.overdrive~",(method)overdrive_new, (method)overdrive_free, (short)sizeof(t_overdrive), 
+	c = class_new("tt.overdrive~",(method)overdrive_new, (method)overdrive_free, sizeof(t_overdrive), 
 		(method)0L, A_GIMME, 0);
 
 	class_addmethod(c, (method)overdrive_setPreamp,				"/preamp",				A_GIMME, 0);
